Enables Dex & Harbor to trust Let's Encrypt (#3)
TL;DR
-----

Updates Dex and Harbor configurations so that deployed pods will
trust Let's Encrypt as a CA

Details
-------

Addreses a challenge I unearthed when using a custom URL (and thus
certificate) for my Okta endpoint. My custom endpoint uses a Let's
Encrypt certificate and the container images used by Dex and Harbor
do not include Let's Encrypt as a trusted certificate. This change
use a YTT overlay to both create a ConfigMap including the Let's
Encrypt CA cert and mount it as a volume so that OpenSSL will see
that certificate as a trusted one.